Npm Package Restructure
In contrast to the bower package the idea is that if you wish to consume SexyTable
in your TypeScript project you would use the npm package.

The npm package will now install transparency as a post install script.
This is because one of it's dependencies fails to build (on my windows system at least).
We only need the compiled transparency source so this is okay.

We also install tsd globally (maybe this is a bit much) and run tsd reinstall.
So that all our TypeScript types get installed for us as well.