Massive Updates to Editor
Instead of requiring the user to double click on the cell to edit it.
Now we wrap all cells that can be edited with an input field.
I feel this is much better from a UI perspective, now it is immediately obvious the cells can be edited.
Thus we no longer need the EditPrompt which I quickly found needed to be adjusted in another project.